Remake a bunch of stuff, multi-level polymorphism and such..

+++ brutalfighters-server/src/com/brutalfighters/server/util/CollisionDetection.java
@@ -17,7 +17,7 @@ public class CollisionDetection {
 	}
 	
 	public static Rectangle getBounds(String flip, float x, float y, float width, float height) {
-		return getBounds(flip, (int)x, (int)y, width, height);
+		return getBounds(flip, (int)x, (int)y, (int)width, (int)height);
 	}
 	
 	public static void setBounds(Rectangle rectangle, String flip, int x, int y, int width, int height) {
@@ -29,7 +29,7 @@ public class CollisionDetection {
 			rectangle.setBounds(x-width/2,y-height/2,width,height);
 		}
 	}
-	public static void setBounds(Rectangle rectangle, String flip, float x, float y, int width, int height) {
-		setBounds(rectangle, flip, (int)x, (int)y, width, height);
+	public static void setBounds(Rectangle rectangle, String flip, float x, float y, float width, float height) {
+		setBounds(rectangle, flip, (int)x, (int)y, (int)width, (int)height);
 	}
 }
